MHD-013 Refactor code to avoid all warnings

+++ src/openGL/IndexBuffer.cpp
@@ -2,9 +2,9 @@
 
 #include "Renderer.h"
 
-IndexBuffer::IndexBuffer(const unsigned int* data, unsigned int count)
+IndexBuffer::IndexBuffer(const unsigned int* data, size_t count)
     : m_Count(count) {
-    ASSERT(sizeof(unsigned int) == sizeof(GLuint));
+    ASSERT(sizeof(size_t) == sizeof(GLuint));
 
     GLCall(glGenBuffers(1, &m_RendererID));
     GLCall(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID));
